agregado metodo getAllStickiesService, cambios en manejo de errores
